refactor: avoid adding `typing.CompliantExpr`
https://github.com/narwhals-dev/narwhals/pull/2149#discussion_r1981760892
