+++ narwhals/typing.py
@@ -8,7 +8,6 @@ from typing import TypeVar
 from typing import Union
 
 from narwhals._compliant import CompliantDataFrame
-from narwhals._compliant import CompliantExpr
 from narwhals._compliant import CompliantLazyFrame
 from narwhals._compliant import CompliantSeries
 
@@ -235,7 +234,6 @@ class DTypes:
 
 __all__ = [
     "CompliantDataFrame",
-    "CompliantExpr",
     "CompliantLazyFrame",
     "CompliantSeries",
     "DataFrameT",
